Don't store size as int in SameSizeSet
This attribute is directly assigned to things like PartitionDevice.req_max_size
which should be an instance of Size. Also a general rule of thumb is to store
sizes as instances of Size so we should do it here too.

I don't really know why we ever converted the value to int() in that place, but
tests seem to pass just fine without the conversion so I guess everything's just
fine.

+++ blivet/partitioning.py
@@ -1584,7 +1584,7 @@ class SameSizeSet(object):
 
             self.devices.append(partition)
 
-        self.size = int(size / len(devices))
+        self.size = size / len(devices)
         self.grow = grow
         self.max_size = max_size
